+++ src/plugins/index.js
@@ -1,6 +1,7 @@
 // #if (!process.env.WEBPACK_BUILD_LIGHT)
 import 'videojs-contrib-ads';
 import dashPlugin from './dash';
+import imaPlugin from './ima';
 import interactive from './interactive-plugin';
 import './videojs-http-source-selector/plugin';
 // #endif
@@ -25,7 +26,8 @@ const plugins = {
   vttThumbnails,
   // #if (!process.env.WEBPACK_BUILD_LIGHT)
   interactive,
-  dashPlugin
+  dashPlugin,
+  imaPlugin
   // #endif
 };
